Fix CI: benchmark gate, ty check, and formatting
- benchmarks/cases.py: misaligned_matmat_vs_scipy/misaligned_rmatmat_vs_scipy
  still called peak_alloc_mb() after #56 dropped that helper entirely from
  benchmarks/harness.py, so every benchmark run failed with NameError. This
  was already broken on main (confirmed: main's own CI is red for the same
  reason), not something introduced by this branch, but it blocks this PR's
  gate too. Drops the two dead peak_alloc_mb entries (case output and
  baselines.json), matching how #56 already removed the other five
  memory-metric cases.
- _ivcsc.py/_rapid_load.py: the shared _decode_varint helper from an earlier
  commit here expects a np.int64 position, but both callers seeded `pos` as
  a plain `0`, which `ty check` flagged as a real (if numba-tolerated) type
  mismatch. Seed `pos = np.int64(0)` instead.
- _rapid_load.py: the _read_ivcsr_group extraction from the same earlier
  commit gives `kwargs` an explicit `dict[str, Any]` return type, which
  resolves the invalid-argument-type ty previously needed ignore comments
  for on `ad.AnnData(**kwargs)` -- removed the two now-unused ignores.
- ruff format on the three files touched by the complexity refactor.

uv run pytest: 1352 passed, 61 skipped. uv run python -m benchmarks.run
--set fast: no regressions. ruff check/format, ty check, codespell all pass.

+++ src/vsparse/_ivcsc.py
@@ -95,7 +95,7 @@ def _decode_varint(buf: np.ndarray, pos: np.int64) -> tuple[np.uint64, np.int64]
 @numba.njit(cache=True)
 def _unpack(value_ptr: np.ndarray, buf: np.ndarray, out: np.ndarray) -> None:
     n_groups = value_ptr.shape[0] - 1
-    pos = 0
+    pos = np.int64(0)
     for g in range(n_groups):
         start, end = value_ptr[g], value_ptr[g + 1]
         prev = np.int64(-1)
